refactor: implementar serviços de autenticação e orquestração de análises, substituindo MediatR
