+++ QuiosqueBI.API/Services/Auth/IAuthService.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Identity;
+
+namespace QuiosqueBI.API.Services;
+
+public interface IAuthService
+{
+    Task<IActionResult> RegisterAsync(string nome, string sobrenome, string email, string password);
+    Task<IActionResult> LoginAsync(string email, string password);
+}
